Ref: Utils.
