+++ src/Utils.ts
@@ -17,170 +17,166 @@ let EXTENSION_NAME: string;
 let EXTENSION_VERSION: string;
 let EXTENSION_AI_KEY: string;
 
-export module Utils {
-    export async function loadPackageInfo(context: vscode.ExtensionContext): Promise<void> {
-        const { publisher, name, version, aiKey } = await fse.readJSON(context.asAbsolutePath("./package.json"));
-        EXTENSION_AI_KEY = aiKey;
-        EXTENSION_PUBLISHER = publisher;
-        EXTENSION_NAME = name;
-        EXTENSION_VERSION = version;
+export async function loadPackageInfo(context: vscode.ExtensionContext): Promise<void> {
+    const { publisher, name, version, aiKey } = await fse.readJSON(context.asAbsolutePath("./package.json"));
+    EXTENSION_AI_KEY = aiKey;
+    EXTENSION_PUBLISHER = publisher;
+    EXTENSION_NAME = name;
+    EXTENSION_VERSION = version;
+}
+
+export function getExtensionId(): string {
+    return `${EXTENSION_PUBLISHER}.${EXTENSION_NAME}`;
+}
+export function getVersion(): string {
+    return EXTENSION_VERSION;
+}
+export function getAiKey(): string {
+    return EXTENSION_AI_KEY;
+}
+export function getTempFolder(): string {
+    return path.join(os.tmpdir(), getExtensionId());
+}
+
+export async function downloadFile(targetUrl: string, readContent?: boolean, customHeaders?: {}): Promise<string> {
+    const tempFilePath: string = path.join(getTempFolder(), md5(targetUrl));
+    await fse.ensureDir(getTempFolder());
+    if (await fse.pathExists(tempFilePath)) {
+        await fse.remove(tempFilePath);
     }
 
-    export function getExtensionId(): string {
-        return `${EXTENSION_PUBLISHER}.${EXTENSION_NAME}`;
-    }
-    export function getVersion(): string {
-        return EXTENSION_VERSION;
-    }
-    export function getAiKey(): string {
-        return EXTENSION_AI_KEY;
-    }
-    export function getTempFolder(): string {
-        return path.join(os.tmpdir(), getExtensionId());
-    }
-
-    export async function downloadFile(targetUrl: string, readContent?: boolean, customHeaders?: {}): Promise<string> {
-        const tempFilePath: string = path.join(getTempFolder(), md5(targetUrl));
-        await fse.ensureDir(getTempFolder());
-        if (await fse.pathExists(tempFilePath)) {
-            await fse.remove(tempFilePath);
-        }
-
-        return await new Promise((resolve: (res: string) => void, reject: (e: Error) => void): void => {
-            const urlObj: url.Url = url.parse(targetUrl);
-            const options: Object = Object.assign({ headers: Object.assign({}, customHeaders, { 'User-Agent': `vscode/${getVersion()}` }) }, urlObj);
-            let client: any;
-            if (urlObj.protocol === "https:") {
-                client = https;
-                // tslint:disable-next-line:no-http-string
-            } else if (urlObj.protocol === "http:") {
-                client = http;
-            } else {
-                return reject(new Error("Unsupported protocol."));
-            }
-            client.get(options, (res: http.IncomingMessage) => {
-                let rawData: string;
-                let ws: fse.WriteStream;
-                if (readContent) {
-                    rawData = "";
-                } else {
-                    ws = fse.createWriteStream(tempFilePath);
-                }
-                res.on('data', (chunk: string | Buffer) => {
-                    if (readContent) {
-                        rawData += chunk;
-                    } else {
-                        ws.write(chunk);
-                    }
-                });
-                res.on('end', () => {
-                    if (readContent) {
-                        resolve(rawData);
-                    } else {
-                        ws.end();
-                        resolve(tempFilePath);
-                    }
-                });
-            }).on("error", (err: Error) => {
-                reject(err);
-            });
-        });
-    }
-
-    export async function writeFileToExtensionRoot(relateivePath: string, data: string | Buffer): Promise<void> {
-        const extensionRootPath: string = vscode.extensions.getExtension(getExtensionId()).extensionPath;
-        const filepath: string = path.join(extensionRootPath, relateivePath);
-        await fse.ensureFile(filepath);
-        await fse.writeFile(filepath, data);
-    }
-
-    export async function readFileFromExtensionRoot(relateivePath: string): Promise<string> {
-        const extensionRootPath: string = vscode.extensions.getExtension(getExtensionId()).extensionPath;
-        const filepath: string = path.join(extensionRootPath, relateivePath);
-        if (await fse.pathExists(filepath)) {
-            const buf: Buffer = await fse.readFile(filepath);
-            return buf.toString();
+    return await new Promise((resolve: (res: string) => void, reject: (e: Error) => void): void => {
+        const urlObj: url.Url = url.parse(targetUrl);
+        const options: Object = Object.assign({ headers: Object.assign({}, customHeaders, { 'User-Agent': `vscode/${getVersion()}` }) }, urlObj);
+        let client: any;
+        if (urlObj.protocol === "https:") {
+            client = https;
+            // tslint:disable-next-line:no-http-string
+        } else if (urlObj.protocol === "http:") {
+            client = http;
         } else {
-            return null;
+            return reject(new Error("Unsupported protocol."));
         }
-    }
-
-    export function groupIdValidation(value: string): string {
-        return (/^[a-z_][a-z0-9_]*(\.[a-z_][a-z0-9_]*)*$/.test(value)) ? null : "Invalid Group Id";
-    }
-
-    export function artifactIdValidation(value: string): string {
-        return (/^[a-z_][a-z0-9_]*(-[a-z_][a-z0-9_]*)*$/.test(value)) ? null : "Invalid Artifact Id";
-    }
-
-    export async function readXmlContent(xml: string, options?: {}): Promise<any> {
-        const opts: {} = Object.assign({ explicitArray: true }, options);
-        return new Promise<{}>(
-            (resolve: (value: {}) => void, reject: (e: Error) => void): void => {
-                xml2js.parseString(xml, opts, (err: Error, res: any) => {
-                    if (err) {
-                        reject(err);
-                    } else {
-                        resolve(res);
-                    }
-                });
-            }
-        );
-    }
-
-    export function buildXmlContent(obj: any, options?: {}): string {
-        const opts: {} = Object.assign({ explicitArray: true }, options);
-        return new xml2js.Builder(opts).buildObject(obj);
-    }
-
-    export namespace settings {
-        export function getServiceUrl(): string {
-            return vscode.workspace.getConfiguration("spring.initializr").get<string>("serviceUrl");
-        }
-    }
-
-    export async function getTargetPomXml(): Promise<vscode.Uri> {
-        if (vscode.window.activeTextEditor) {
-            const fsPath: string = vscode.window.activeTextEditor.document.fileName;
-            if ("pom.xml" === path.basename(fsPath).toLowerCase()) {
-                return vscode.Uri.file(fsPath);
-            }
-        }
-
-        const candidates: vscode.Uri[] = await vscode.workspace.findFiles('**/pom.xml');
-        if (!_.isEmpty(candidates)) {
-            if (candidates.length === 1) {
-                return candidates[0];
+        client.get(options, (res: http.IncomingMessage) => {
+            let rawData: string;
+            let ws: fse.WriteStream;
+            if (readContent) {
+                rawData = "";
             } else {
-                return await VSCodeUI.getQuickPick(
-                    candidates,
-                    getRelativePathToWorkspaceFolder,
-                    getWorkspaceFolderName,
-                    null,
-                    { placeHolder: "Select the target project."}
-                );
+                ws = fse.createWriteStream(tempFilePath);
             }
+            res.on('data', (chunk: string | Buffer) => {
+                if (readContent) {
+                    rawData += chunk;
+                } else {
+                    ws.write(chunk);
+                }
+            });
+            res.on('end', () => {
+                if (readContent) {
+                    resolve(rawData);
+                } else {
+                    ws.end();
+                    resolve(tempFilePath);
+                }
+            });
+        }).on("error", (err: Error) => {
+            reject(err);
+        });
+    });
+}
+
+export async function writeFileToExtensionRoot(relateivePath: string, data: string | Buffer): Promise<void> {
+    const extensionRootPath: string = vscode.extensions.getExtension(getExtensionId()).extensionPath;
+    const filepath: string = path.join(extensionRootPath, relateivePath);
+    await fse.ensureFile(filepath);
+    await fse.writeFile(filepath, data);
+}
+
+export async function readFileFromExtensionRoot(relateivePath: string): Promise<string> {
+    const extensionRootPath: string = vscode.extensions.getExtension(getExtensionId()).extensionPath;
+    const filepath: string = path.join(extensionRootPath, relateivePath);
+    if (await fse.pathExists(filepath)) {
+        const buf: Buffer = await fse.readFile(filepath);
+        return buf.toString();
+    } else {
+        return null;
+    }
+}
+
+export function groupIdValidation(value: string): string {
+    return (/^[a-z_][a-z0-9_]*(\.[a-z_][a-z0-9_]*)*$/.test(value)) ? null : "Invalid Group Id";
+}
+
+export function artifactIdValidation(value: string): string {
+    return (/^[a-z_][a-z0-9_]*(-[a-z_][a-z0-9_]*)*$/.test(value)) ? null : "Invalid Artifact Id";
+}
+
+export async function readXmlContent(xml: string, options?: {}): Promise<any> {
+    const opts: {} = Object.assign({ explicitArray: true }, options);
+    return new Promise<{}>(
+        (resolve: (value: {}) => void, reject: (e: Error) => void): void => {
+            xml2js.parseString(xml, opts, (err: Error, res: any) => {
+                if (err) {
+                    reject(err);
+                } else {
+                    resolve(res);
+                }
+            });
         }
-        return undefined;
+    );
+}
+
+export function buildXmlContent(obj: any, options?: {}): string {
+    const opts: {} = Object.assign({ explicitArray: true }, options);
+    return new xml2js.Builder(opts).buildObject(obj);
+}
+
+export function getServiceUrl(): string {
+    return vscode.workspace.getConfiguration("spring.initializr").get<string>("serviceUrl");
+}
+
+export async function getTargetPomXml(): Promise<vscode.Uri> {
+    if (vscode.window.activeTextEditor) {
+        const fsPath: string = vscode.window.activeTextEditor.document.fileName;
+        if ("pom.xml" === path.basename(fsPath).toLowerCase()) {
+            return vscode.Uri.file(fsPath);
+        }
     }
 
-    function getRelativePathToWorkspaceFolder(file: vscode.Uri): string {
-        if (file) {
-            const wf: vscode.WorkspaceFolder = vscode.workspace.getWorkspaceFolder(file);
-            if (wf) {
-                return path.relative(wf.uri.fsPath, file.fsPath);
-            }
+    const candidates: vscode.Uri[] = await vscode.workspace.findFiles('**/pom.xml');
+    if (!_.isEmpty(candidates)) {
+        if (candidates.length === 1) {
+            return candidates[0];
+        } else {
+            return await VSCodeUI.getQuickPick(
+                candidates,
+                getRelativePathToWorkspaceFolder,
+                getWorkspaceFolderName,
+                null,
+                { placeHolder: "Select the target project." }
+            );
         }
-        return "";
     }
+    return undefined;
+}
 
-    function getWorkspaceFolderName(file: vscode.Uri): string {
-        if (file) {
-            const wf: vscode.WorkspaceFolder = vscode.workspace.getWorkspaceFolder(file);
-            if (wf) {
-                return wf.name;
-            }
+function getRelativePathToWorkspaceFolder(file: vscode.Uri): string {
+    if (file) {
+        const wf: vscode.WorkspaceFolder = vscode.workspace.getWorkspaceFolder(file);
+        if (wf) {
+            return path.relative(wf.uri.fsPath, file.fsPath);
         }
-        return "";
     }
+    return "";
+}
+
+function getWorkspaceFolderName(file: vscode.Uri): string {
+    if (file) {
+        const wf: vscode.WorkspaceFolder = vscode.workspace.getWorkspaceFolder(file);
+        if (wf) {
+            return wf.name;
+        }
+    }
+    return "";
 }
